python/requirements.txt
web3>=6.20.2
python-dotenv>=1.0.1
eth-account>=0.13.0
